requirements.txt
Django>=4.0.4,<4.1.0
djangorestframework>=3.13.1,<3.14.0
django-cors-headers

django-allauth
dj-rest-auth
djangorestframework-simplejwt
# django-sendgrid-v5


psycopg2-binary==2.9.3

flake8>=4.0.1,<4.1.0